fix : 안티패턴 픽스~~~~~예이
