+++ src/components/pages/ticketList/BigPoster.jsx
@@ -7,7 +7,7 @@ import useItems from '@hooks/useItems';
 const BigPoster = ({ type, items }) => {
   const [activePosterId, setActivePosterId] = useState(0);
 
-  const { typeItems } = useItems({ type, items });
+  const { typeItems } = useItems();
 
   const onClickCard = (side) => {
     if (activePosterId === 0 && side === 'right') {
